ML-SAST/requirements.txt
decorator==4.4.2
docker==5.0.3
GitPython==3.1.27
numpy==1.23.0
pandas==1.3.5
pdoc==11.0.0
prompt-toolkit==3.0.31
pytest==7.1.1
PyYAML==6.0
requests==2.27.1
scikit-learn==1.1.1
swifter==1.3.4
